[CORE-10981] Move bandwidth code to pkg/bandwidth
Move bandwidth functionality to pkg/bandwidth, make it an importable
module. Import it and use it in plugins/meta/bandwidth/main.go.

Add a qdiscType arg to CreateEgressQdisc() so that both "ingress" and
"clsact" qdiscs may be used ("ingress" is the default used by the
bandwidth plugin).

All of this has the objective of enabling Calico to apply the netlink
qdisc configuration natively by importing this module.